pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["cleiveliu <cleiveliu1@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_template"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2"
django-simpleui = "^2021.6.2"
djangorestframework = "^3.12.4"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
ipython = "^7.24.1"
django-extensions = "^3.1.3"
black = "^21.6-beta.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
